chore: extract indoor layers
